+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [5.0.0-beta.2](https://github.com/maevsi/sqitch/compare/5.0.0-beta.1...5.0.0-beta.2) (2024-12-12)
+
+### ⚠ BREAKING CHANGES
+
+* **notification:** align timestamp column name
+
+### Features
+
+* **notification:** align timestamp column name ([92b2ec2](https://github.com/maevsi/sqitch/commit/92b2ec28752a96b70d4a51256959032074259b6d))
+
+## [5.0.0-beta.1](https://github.com/maevsi/sqitch/compare/4.3.1...5.0.0-beta.1) (2024-12-12)
+
+### ⚠ BREAKING CHANGES
+
+* **account:** rename `created` column to `created_at`
+
+### Features
+
+* **account:** rename `created` column to `created_at` ([df18548](https://github.com/maevsi/sqitch/commit/df18548e11871a22271fa9d131a0538782c1e51b))
+
 ## [4.9.1](https://github.com/maevsi/sqitch/compare/4.9.0...4.9.1) (2024-12-18)
 
 ### Bug Fixes